requirements.txt
pyeodh
python-dotenv
click